requirements.txt
nest_asyncio==1.6.0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2022.1
